+++ beets/autotag/mb.py
@@ -188,9 +188,16 @@ def album_info(release):
     info.va = info.artist_id == VARIOUS_ARTISTS_ID
     info.asin = release.get('asin')
     info.releasegroup_id = release['release-group']['id']
-    info.albumdisambig = release['release-group'].get('disambiguation')
     info.country = release.get('country')
     info.albumstatus = release.get('status')
+
+    # Build up the disambiguation string from the release group and release.
+    disambig = []
+    if release['release-group'].get('disambiguation'):
+        disambig.append(release['release-group'].get('disambiguation'))
+    if release.get('disambiguation'):
+        disambig.append(release.get('disambiguation'))
+    info.albumdisambig = u', '.join(disambig)
 
     # Release type not always populated.
     if 'type' in release['release-group']: