+++ test/test_mb.py
@@ -23,12 +23,12 @@ class MBAlbumInfoTest(unittest.TestCase):
             'title': 'ALBUM TITLE',
             'id': 'ALBUM ID',
             'asin': 'ALBUM ASIN',
-            'disambiguation': 'DISAMBIGUATION',
+            'disambiguation': 'R_DISAMBIGUATION',
             'release-group': {
                 'type': 'Album',
                 'first-release-date': date_str,
                 'id': 'RELEASE GROUP ID',
-                'disambiguation': 'DISAMBIGUATION',
+                'disambiguation': 'RG_DISAMBIGUATION',
             },
             'artist-credit': [
                 {
@@ -248,7 +248,8 @@ class MBAlbumInfoTest(unittest.TestCase):
     def test_parse_disambig(self):
         release = self._make_release(None)
         d = mb.album_info(release)
-        self.assertEqual(d.albumdisambig, 'DISAMBIGUATION')
+        self.assertEqual(d.albumdisambig,
+                         'RG_DISAMBIGUATION, R_DISAMBIGUATION')
 
     def test_parse_disctitle(self):
         tracks = [self._make_track('TITLE ONE', 'ID ONE', 100.0 * 1000.0),